requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==4.2.6
djangorestframework==3.14.0
packaging==23.2
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
